requirements.txt
fastapi
uvicorn
python-dotenv
line-bot-sdk
discord-interactions
openai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
